requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.1.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
debugpy==1.6.3
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
flatbuffers==22.10.26
gast==0.4.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.4
jupyter_core==4.11.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
termcolor==2.1.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
gunicorn==20.1.0
